agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "livekit-agents[openai]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
]
